refactor bootstrap ci
