+++ examples/time_frequency/plot_time_frequency_global_field_power.py
@@ -46,6 +46,8 @@ import matplotlib.pyplot as plt
 
 import mne
 from mne.datasets import somato
+from mne.baseline import rescale
+from mne.utils import _bootstrap_ci
 
 ###############################################################################
 # Set parameters
@@ -96,27 +98,8 @@ for band, fmin, fmax in iter_freqs:
 
 ###############################################################################
 # Now we can compute the Global Field Power
-#
-# Then we prepare a bootstrapping function to estimate confidence intervals
-
-rng = np.random.RandomState(42)
-
-
-def get_gfp_ci(average, n_bootstraps=2000):
-    """get confidence intervals from non-parametric bootstrap"""
-    indices = np.arange(len(average.ch_names), dtype=int)
-    gfps_bs = np.empty((n_bootstraps, len(average.times)))
-    for iteration in range(n_bootstraps):
-        bs_indices = rng.choice(indices, replace=True, size=len(indices))
-        gfps_bs[iteration] = np.sum(average.data[bs_indices] ** 2, 0)
-    gfps_bs = mne.baseline.rescale(gfps_bs, average.times, baseline=(None, 0))
-    ci_low, ci_up = np.percentile(gfps_bs, (2.5, 97.5), axis=0)
-    return ci_low, ci_up
-
-
-##############################################################################
-# Now we can track the emergence of spatial patterns compared to baseline
-# for each frequency band
+# We can track the emergence of spatial patterns compared to baseline
+# for each frequency band, with a bootstrapped confidence interval.
 #
 # We see dominant responses in the Alpha and Beta bands.
 
@@ -128,17 +111,18 @@ for ((freq_name, fmin, fmax), average), color, ax in zip(
     gfp = np.sum(average.data ** 2, axis=0)
     gfp = mne.baseline.rescale(gfp, times, baseline=(None, 0))
     ax.plot(times, gfp, label=freq_name, color=color, linewidth=2.5)
-    ax.plot(times, np.zeros_like(times), linestyle='--', color='red',
-            linewidth=1)
-    ci_low, ci_up = get_gfp_ci(average)
-    ax.fill_between(times, gfp + ci_up, gfp - ci_low, color=color,
-                    alpha=0.3)
+    ax.axhline(0, linestyle='--', color='grey', linewidth=2)
+    ci_low, ci_up = _bootstrap_ci(average.data, random_state=0,
+                                  statfun=lambda x: np.sum(x ** 2, axis=0))
+    ci_low = rescale(ci_low, average.times, baseline=(None, 0))
+    ci_up = rescale(ci_up, average.times, baseline=(None, 0))
+    ax.fill_between(times, gfp + ci_up, gfp - ci_low, color=color, alpha=0.3)
     ax.grid(True)
     ax.set_ylabel('GFP')
     ax.annotate('%s (%d-%dHz)' % (freq_name, fmin, fmax),
                 xy=(0.95, 0.8),
                 horizontalalignment='right',
                 xycoords='axes fraction')
-    ax.set_xlim(-1050, 3050)
+    ax.set_xlim(-1000, 3000)
 
 axes.ravel()[-1].set_xlabel('Time [ms]')
